Update path to moved ensureLoaders function
